Release 1.2.0
I’m bumping to a new minor version after an extensive rewrite of the
undocumented VCR Matcher, trying to handle some edge cases when testing
posts with bodies. This is backwards-incompatible, but I’m not doing a
major version bump because the class is undocumented and I still
consider it private.